+++ lib/models/model.dart
@@ -10,3 +10,5 @@ export 'presensi/model_token.dart';
 export 'presensi/model_presensi.dart';
 export 'presensi/model_jadwal.dart';
 export 'presensi/model_face.dart';
+export 'presensi/model_event.dart';
+export 'presensi/model_hari_kerja.dart';