+++ lib/screens/beranda/beranda.dart
@@ -1,6 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:absen_online/configs/config.dart';
 import 'package:absen_online/models/model.dart';
+import 'package:pull_to_refresh/pull_to_refresh.dart';
 import 'beranda_sliver_app_bar.dart';
 import 'package:absen_online/utils/utils.dart';
 import 'package:absen_online/blocs/bloc.dart';
@@ -17,6 +18,7 @@ class Beranda extends StatefulWidget {
 
 class _BerandaState extends State<Beranda> {
   JadwalCubit _jadwalCubit;
+  final _controller = RefreshController(initialRefresh: false);
 
   @override
   void initState() {
@@ -63,7 +65,9 @@ class _BerandaState extends State<Beranda> {
         return AppInfo(
           title: state.info['title'].toString(),
           message: state.info['content'].toString(),
-          image: state.info['type'].toString() == 'hari' ? Images.Calendar : Images.Calendar,
+          image: state.info['type'].toString() == 'hari'
+              ? Images.Calendar
+              : Images.Calendar,
         );
       } else if (state is JadwalError) {
         return AppError(
@@ -87,62 +91,99 @@ class _BerandaState extends State<Beranda> {
     });
   }
 
+  Future<void> _onRefresh() async {
+    _jadwalCubit.reInit();
+    _jadwalCubit.initData();
+    _controller.refreshCompleted();
+  }
+
   @override
   Widget build(BuildContext context) {
     return Scaffold(
-      body: CustomScrollView(
-        slivers: <Widget>[
-          SliverPersistentHeader(
-            delegate: AppBarHomeSliver(
-                expandedHeight: 300,
-                banners: Application.remoteConfig?.banner == null
-                    ? []
-                    : Application.remoteConfig.banner
-                        .map((banner) => ImageModel(1, banner))
-                        .toList()),
-            pinned: true,
+      body: SmartRefresher(
+        enablePullDown: true,
+        enablePullUp: false,
+        onRefresh: _onRefresh,
+        // onLoading: _onLoading,
+        controller: _controller,
+        header: ClassicHeader(
+          idleText: Translate.of(context).translate('pull_down_refresh'),
+          refreshingText: Translate.of(context).translate('refreshing'),
+          completeText: Translate.of(context).translate('refresh_completed'),
+          releaseText: Translate.of(context).translate('release_to_refresh'),
+          refreshingIcon: SizedBox(
+            width: 16.0,
+            height: 16.0,
+            child: CircularProgressIndicator(strokeWidth: 2),
           ),
-          SliverList(
-            delegate: SliverChildListDelegate([
-              SafeArea(
-                top: false,
-                bottom: false,
-                child: Column(
-                  children: <Widget>[
-                    Container(
-                      padding: EdgeInsets.only(
-                        left: 20,
-                        right: 20,
-                        bottom: 15,
+        ),
+        footer: ClassicFooter(
+          loadingText: Translate.of(context).translate('loading'),
+          canLoadingText: Translate.of(context).translate(
+            'release_to_load_more',
+          ),
+          idleText: Translate.of(context).translate('pull_to_load_more'),
+          loadStyle: LoadStyle.ShowWhenLoading,
+          loadingIcon: SizedBox(
+            width: 16.0,
+            height: 16.0,
+            child: CircularProgressIndicator(strokeWidth: 2),
+          ),
+        ),
+        child: CustomScrollView(
+          slivers: <Widget>[
+            SliverPersistentHeader(
+              delegate: AppBarHomeSliver(
+                  expandedHeight: 300,
+                  banners: Application.remoteConfig?.banner == null
+                      ? []
+                      : Application.remoteConfig.banner
+                          .map((banner) => ImageModel(1, banner))
+                          .toList()),
+              pinned: true,
+            ),
+            SliverList(
+              delegate: SliverChildListDelegate([
+                SafeArea(
+                  top: false,
+                  bottom: false,
+                  child: Column(
+                    children: <Widget>[
+                      Container(
+                        padding: EdgeInsets.only(
+                          left: 20,
+                          right: 20,
+                          bottom: 15,
+                        ),
+                        child: Row(
+                          children: <Widget>[
+                            Column(
+                              crossAxisAlignment: CrossAxisAlignment.start,
+                              children: <Widget>[
+                                Text(
+                                  Translate.of(context)
+                                      .translate('attendance_schedule'),
+                                  style: Theme.of(context)
+                                      .textTheme
+                                      .headline6
+                                      .copyWith(fontWeight: FontWeight.w600),
+                                ),
+                              ],
+                            )
+                          ],
+                        ),
                       ),
-                      child: Row(
-                        children: <Widget>[
-                          Column(
-                            crossAxisAlignment: CrossAxisAlignment.start,
-                            children: <Widget>[
-                              Text(
-                                Translate.of(context)
-                                    .translate('attendance_schedule'),
-                                style: Theme.of(context)
-                                    .textTheme
-                                    .headline6
-                                    .copyWith(fontWeight: FontWeight.w600),
-                              ),
-                            ],
-                          )
-                        ],
+                      Container(
+                        padding: EdgeInsets.only(left: 20, right: 20),
+                        child: _buildListJadwal(),
                       ),
-                    ),
-                    Container(
-                      padding: EdgeInsets.only(left: 20, right: 20),
-                      child: _buildListJadwal(),
-                    ),
-                  ],
-                ),
-              )
-            ]),
-          )
-        ],
+                    ],
+                  ),
+                )
+              ]),
+            )
+          ],
+        ),
       ),
     );
   }